--- src/docfx/lib/attribute/MarkdownAttribute.cs
@@ -1,15 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.Docs.Build
-{
-
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class MarkdownAttribute : SchemaContentTypeAttribute
-    {
-        public MarkdownAttribute()
-            : base(SchemaContentType.Markdown) { }
-    }
-}
